Avoid some potential namespace conflicts.
`pg-data` is a common name for postgres docker volumes. A volume with this name may already exist on a users system which will produce an error message when trying to run `docker compose up`

I've also prefixed the spark images with `anaml`. This means when a user does `docker ps` all anaml images are grouped together.